Check return types

+++ tests/xy_c_compile_resources/positionalTags.c
@@ -13,7 +13,7 @@ struct positionalTags_Point {
     float m_y;
 };
 
-int32_t positionalTags_test(void) {
+size_t positionalTags_test(void) {
     positionalTags_Array ints = (positionalTags_Array){0};
     positionalTags_Array points = (positionalTags_Array){0};
     return ints.m_len + points.m_len + 4 + 8;
